11
111

+++ src/main/java/com/choong/spr/controller/MemberController.java
@@ -3,6 +3,8 @@ package com.choong.spr.controller;
 import java.security.Principal;
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -89,9 +91,9 @@ public class MemberController {
 	}
 
 	@GetMapping("get")
-	public String getMember(String id, Model model,Principal principal) {
+	public String getMember(String id, Model model,Principal principal, HttpServletRequest request) {
 		
-		if(principal !=null && principal.getName().equals(id)) {
+		if(hasAuthOrAdmin(id, principal,request)) {
 				MemberDto member = service.getMemberById(id);
 				model.addAttribute("member", member);	
 				
@@ -100,11 +102,16 @@ public class MemberController {
 		return "redirect:/member/login";
 		
 	}
-
+	private boolean hasAuthOrAdmin(String id,Principal principal,HttpServletRequest req) {
+		return req.isUserInRole("ROLE_ADMIN") || (principal !=null && principal.getName().equals(id));
+	}
+	
 	@PostMapping("remove")
-	public String removeMember(MemberDto dto, RedirectAttributes rttr) {
-		boolean success = service.removeMember(dto);
-
+	public String removeMember(MemberDto dto, RedirectAttributes rttr, Principal principal, HttpServletRequest req) {
+			
+		if (hasAuthOrAdmin(dto.getId(), principal, req)) {
+			boolean success = service.removeMember(dto);
+		
 		if (success) {
 			rttr.addFlashAttribute("message", "회원 탈퇴 되었습니다.");
 			return "redirect:/board/list";
@@ -112,12 +119,16 @@ public class MemberController {
 			rttr.addFlashAttribute("id", dto.getId());
 			return "redirect:/member/get";
 		}
+		}else {
+			return "redirect:/member/login";
+		}
 
 	}
 
 	@PostMapping("modify")
-	public String modifyMember(MemberDto dto, String oldPassword, RedirectAttributes rttr) {
-
+	public String modifyMember(MemberDto dto, String oldPassword,Principal principal,
+			HttpServletRequest req, RedirectAttributes rttr) {
+		if (hasAuthOrAdmin(dto.getId(), principal, req)) {
 		boolean success = service.modifyMember(dto, oldPassword);
 		if (success) {
 			rttr.addFlashAttribute("message", "회원정보가 수정되었다");
@@ -128,11 +139,23 @@ public class MemberController {
 		rttr.addAttribute("id", dto.getId());
 
 		return "redirect:/member/get";
+	}else {
+		return "redirect:/member/login";
 	}
-
+	}
 	@GetMapping("login")
 	public void loginPage() {
 
 	}
-
+	
+	@GetMapping("initpw")
+	public void initpwPage() {
+		
+	}
+	
+	@PostMapping("initpw")
+	public String initPassword(String id) {
+		service.initPassword(id);
+		return "redirect:/board/list";
+		}
 }
